Grant batt_health proper unix perms
There's no need to run batt_health as root, mot_pwric is enough as
long as we fix the permissions of the 'capacity' and 'status'.

The service would also need to be in the logd group to chmod
/data/power_supply_logger/power_supply_info.bin, but nothing uses
this file currently, so it's not required.

Change-Id: I06c33cd05dbc7ab276619400f974c9682758e9e0Grant batt_health proper unix perms

There's no need to run batt_health as root, mot_pwric is enough as
long as we fix the permissions of the 'capacity' and 'status'.

The service would also need to be in the logd group to chmod
/data/power_supply_logger/power_supply_info.bin, but nothing uses
this file currently, so it's not required.

Change-Id: I06c33cd05dbc7ab276619400f974c9682758e9e0

Signed-off-by: klozz <shaclone707@gmail.com>